+++ modules/carto/src/api/layer-map.ts
@@ -11,9 +11,11 @@ import {
 } from 'd3-scale';
 import {format as d3Format} from 'd3-format';
 import moment from 'moment-timezone';
-import {log} from '@deck.gl/core';
 import {H3HexagonLayer, MVTLayer} from '@deck.gl/geo-layers';
 import {GeoJsonLayer} from '@deck.gl/layers';
+import CartoTileLayer from '../layers/carto-tile-layer';
+import {TILE_FORMATS} from './maps-api-common';
+import {assert} from '../utils';
 
 const SCALE_FUNCS = {
   linear: scaleLinear,
@@ -24,6 +26,7 @@ const SCALE_FUNCS = {
   quantize: scaleQuantize,
   sqrt: scaleSqrt
 };
+export type SCALE_TYPE = keyof typeof SCALE_FUNCS;
 
 // Kepler -> Deck.gl
 const sharedPropMap = {
@@ -63,9 +66,16 @@ function mergePropMaps(a, b) {
   return {...a, ...b, visConfig: {...a.visConfig, ...b.visConfig}};
 }
 
-export function getLayer(type, config) {
-  const hexagonId = config.columns?.hex_id;
+export function getLayer(
+  type: string,
+  config,
+  dataset
+): {Layer: any; propMap: any; defaultProps: any} {
+  if (type === 'mvt' || type === 'tileset') {
+    return getTileLayer(dataset);
+  }
 
+  const hexagonId = config.columns?.hex_id;
   const layer = {
     point: {
       Layer: GeoJsonLayer,
@@ -85,25 +95,36 @@ export function getLayer(type, config) {
         visConfig: {coverage: 'coverage', elevationScale: 'elevationScale'}
       }),
       defaultProps: {...defaultProps, getHexagon: d => d[hexagonId]}
-    },
-    mvt: {
-      Layer: MVTLayer,
-      propMap: sharedPropMap,
-      defaultProps: {
-        ...defaultProps,
-        pointRadiusScale: 0.3,
-        lineWidthScale: 2,
-        uniqueIdProperty: 'geoid'
-      }
     }
   }[type];
 
-  log.assert(layer, `Unsupported layer type: ${type}`);
-
+  assert(layer, `Unsupported layer type: ${type}`);
   return layer;
 }
 
-function domainFromAttribute(attribute, scaleType) {
+function getTileLayer(dataset) {
+  const {
+    data: {
+      tiles: [tileUrl]
+    }
+  } = dataset;
+  /* global URL */
+  const formatTiles = new URL(tileUrl).searchParams.get('formatTiles') || TILE_FORMATS.MVT;
+
+  return {
+    Layer: formatTiles === TILE_FORMATS.MVT ? MVTLayer : CartoTileLayer,
+    propMap: sharedPropMap,
+    defaultProps: {
+      ...defaultProps,
+      pointRadiusScale: 0.3,
+      lineWidthScale: 2,
+      uniqueIdProperty: 'geoid',
+      formatTiles
+    }
+  };
+}
+
+function domainFromAttribute(attribute, scaleType: SCALE_TYPE) {
   if (scaleType === 'ordinal' || scaleType === 'point') {
     return attribute.categories.map(c => c.category).filter(c => c !== undefined && c !== null);
   }
@@ -115,13 +136,13 @@ function domainFromAttribute(attribute, scaleType) {
   return [min, attribute.max];
 }
 
-function domainFromValues(values, scaleType) {
+function domainFromValues(values, scaleType: SCALE_TYPE) {
   if (scaleType === 'ordinal') {
     return [...new Set(values)].sort();
   } else if (scaleType === 'quantile') {
     return values.sort((a, b) => a - b);
   } else if (scaleType === 'log') {
-    const [d0, d1] = extent(values);
+    const [d0, d1] = extent(values as number[]);
     return [d0 === 0 ? 1e-5 : d0, d1];
   }
   return extent(values);
@@ -155,8 +176,14 @@ function normalizeAccessor(accessor, data) {
   return accessor;
 }
 
-export function getColorAccessor({name}, scaleType, {colors}, opacity, data) {
-  const scale = SCALE_FUNCS[scaleType]();
+export function getColorAccessor(
+  {name},
+  scaleType: SCALE_TYPE,
+  {colors},
+  opacity: number | undefined,
+  data: any
+) {
+  const scale = SCALE_FUNCS[scaleType as any]();
   scale.domain(calculateDomain(data, name, scaleType));
   scale.range(colors);
   const alpha = opacity !== undefined ? Math.round(255 * Math.pow(opacity, 1 / 2.2)) : 255;
@@ -168,8 +195,8 @@ export function getColorAccessor({name}, scaleType, {colors}, opacity, data) {
   return normalizeAccessor(accessor, data);
 }
 
-export function getSizeAccessor({name}, scaleType, range, data) {
-  const scale = SCALE_FUNCS[scaleType]();
+export function getSizeAccessor({name}, scaleType: SCALE_TYPE, range: Iterable<Range>, data: any) {
+  const scale = SCALE_FUNCS[scaleType as any]();
   scale.domain(calculateDomain(data, name, scaleType));
   scale.range(range);
 
@@ -179,7 +206,7 @@ export function getSizeAccessor({name}, scaleType, range, data) {
   return normalizeAccessor(accessor, data);
 }
 
-const FORMATS = {
+const FORMATS: Record<string, (value: any) => string> = {
   date: s => moment.utc(s).format('MM/DD/YY HH:mm:ssa'),
   integer: d3Format('i'),
   float: d3Format('.5f'),