+++ modules/carto/src/index.ts
@@ -1,12 +1,13 @@
-export {getDefaultCredentials, setDefaultCredentials} from './config.js';
+export {getDefaultCredentials, setDefaultCredentials} from './config';
 export {default as CartoLayer} from './layers/carto-layer';
-export {default as _CartoDynamicTileLayer} from './layers/carto-dynamic-tile-layer';
+export {default as _CartoTileLayer} from './layers/carto-tile-layer';
 export {default as BASEMAP} from './basemap';
 export {default as colorBins} from './style/color-bins-style';
 export {default as colorCategories} from './style/color-categories-style';
 export {default as colorContinuous} from './style/color-continuous-style';
 export {
   FORMATS,
+  TILE_FORMATS,
   MAP_TYPES,
   API_VERSIONS,
   fetchLayerData,