+++ test/modules/carto/api/parseMap.spec.js
@@ -2,6 +2,7 @@ import test from 'tape-catch';
 import VISSTATE_DATA from '../data/visState.json';
 import UNSUPPORTED_LAYER_TYPE_VISSTATE_DATA from '../data/unsupportedLayerTypeVisState.json';
 import {parseMap} from '@deck.gl/carto/api/parseMap';
+import {TILE_FORMATS} from '@deck.gl/carto';
 
 const METADATA = {
   id: 1234,
@@ -30,6 +31,59 @@ const DATASETS = [
   {
     id: 'DATA_TILESET_ID',
     data: {
+      tiles: [
+        `https://gcp-us-east1.api.carto.com/v3/maps/my_connection/tileset/{z}/{x}/{y}?name=my_data&formatTiles=${TILE_FORMATS.MVT}`
+      ],
+      tilestats: {
+        layers: [
+          {
+            attributes: [
+              {
+                attribute: 'STRING_ATTR',
+                categories: [{category: '1'}, {category: '2'}, {category: '3'}]
+              },
+              {
+                attribute: 'NUMBER_ATTR',
+                min: 0,
+                max: 10
+              }
+            ]
+          }
+        ]
+      }
+    }
+  },
+  {
+    id: 'DATA_TILESET_GEOJSON_FORMAT_ID',
+    data: {
+      tiles: [
+        `https://gcp-us-east1.api.carto.com/v3/maps/my_connection/tileset/{z}/{x}/{y}?name=my_data&formatTiles=${TILE_FORMATS.GEOJSON}`
+      ],
+      tilestats: {
+        layers: [
+          {
+            attributes: [
+              {
+                attribute: 'STRING_ATTR',
+                categories: [{category: '1'}, {category: '2'}, {category: '3'}]
+              },
+              {
+                attribute: 'NUMBER_ATTR',
+                min: 0,
+                max: 10
+              }
+            ]
+          }
+        ]
+      }
+    }
+  },
+  {
+    id: 'DATA_TILESET_BINARY_FORMAT_ID',
+    data: {
+      tiles: [
+        `https://gcp-us-east1.api.carto.com/v3/maps/my_connection/tileset/{z}/{x}/{y}?name=my_data&formatTiles=${TILE_FORMATS.BINARY}`
+      ],
       tilestats: {
         layers: [
           {
@@ -108,6 +162,12 @@ for (const {title, visState, layers} of VISSTATE_DATA) {
     layers.forEach(({props: layerProps}) => {
       if (layerProps.data === 'DATA_TILESET') {
         layerProps.data = DATASETS[1].data;
+      }
+      if (layerProps.data === 'DATA_TILESET_GEOJSON') {
+        layerProps.data = DATASETS[2].data;
+      }
+      if (layerProps.data === 'DATA_TILESET_BINARY') {
+        layerProps.data = DATASETS[3].data;
       }
     });