+++ CHANGELOG.md
@@ -21,6 +21,23 @@ Ref: http://keepachangelog.com/en/0.3.0/
 
 ### deck.gl v8.7 Prereleases
 
+#### deck.gl [8.7.0-beta.6] - Feb 21 2022
+- CARTO: Rename CartoDynamicTileLayer -> CartoTileLayer (#6658)
+- GeoJsonLayer[binary]: do not pass instancePickingColors attribute to TextLayer (#6650)
+- Support tilt & bearing for Google raster maps (#6653)
+- Bump ajv from 6.10.2 to 6.12.6 (#6642)
+- Bump follow-redirects from 1.14.7 to 1.14.8 (#6641)
+- Bump simple-get from 3.1.0 to 3.1.1 (#6614)
+#### deck.gl [8.7.0-beta.5] - Feb 10 2022
+
+- CARTO: fetchMap supports tiles in different formats (#6637)
+- CARTO: Simplify format tiles logic (#6635)
+- Carto: Allow private maps in fetchMap (#6629)
+- Change highlightedObjectIndex default to null (#6631)
+#### deck.gl [8.7.0-beta.4] - Feb 7 2022
+
+- Include extensions sent in properties in the CartoDynamicTileLayer
+- Support TILE_FORMATS.BINARY in CartoDynamicTileLayer (#6580)
 #### deck.gl [8.7.0-beta.3] - Feb 2 2022
 
 - CARTO module: Support tiles in different formats (#6609)